Add new feature management classes and modules
This commit implements the new structure for managing features, resources, actions and groups in Featury. The classes and modules include definitions for the DSL used by the application, as well as collections that handle the related objects. Supporting workspace and callable functionality has also been provided. Added new test to ensure correct functionality.