Add group collection to service workflow
The commit introduces a group collection to the service workflows in various files. This update allows group collections to be included as input, handled internally and designed to produce output like other collections. The changes also affect any classes that previously used the modified services, due to the addition of the 'collection_of_groups' attribute. Additionally, outdated resource types in the feature classes are replaced with their updated counterparts.