Add documentation for 'with' method in README
The update in the README.md file introduces the use of the 'with' method for passing arguments in the context of user feature onboarding. An updated ruby code snippet is included in the documentation to illustrate how 'with' method can be used and how it interacts with features' enable and disable functionalities.